.ipynb_checkpoints/requirements-checkpoint.txt
albumentations==0.5.2
gym==0.17.*
pymunk==5.6.0
ml-collections==0.1.0
scikit-learn==0.24.1
tensorflow-cpu
torch==1.13.1
imageio
imageio-ffmpeg
tqdm
torchvision
absl-py
numpy
scipy
pandas
gdown==4.4.0
protobuf~=3.19.0
git+https://github.com/kevinzakka/torchkit.git@v0.0.2
git+https://github.com/kevinzakka/x-magical.git@v0.0.2
